FIX driver path and version

+++ utils.py
@@ -12,6 +12,7 @@ import re
 import time
 import zipfile
 from random import randint
+from sys import platform
 
 from selenium import webdriver
 from selenium.common.exceptions import NoSuchElementException, WebDriverException
@@ -22,19 +23,43 @@ from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.wait import WebDriverWait
 
 
-def download_chrome_driver():
+def get_os_system():
+    if platform == "linux" or platform == "linux2":
+        print("linux")
+    elif platform == "darwin":
+        print("darwin")
+    elif platform == "win32":
+        print("Windows")
 
+
+def download_chrome_driver():
     cwd = os.getcwd()
     folder_driver_path = os.path.join(cwd, "driver")
     driver_path = os.path.join(cwd, "driver", "chromedriver.exe")
+
+    if platform == "linux" or platform == "linux2":
+        chrome_type = "chromedriver_linux64"
+    elif platform == "win32":
+        chrome_type = "chromedriver_win32"
+
     if not os.path.exists(driver_path):
         if not os.path.isdir(folder_driver_path):
             os.mkdir("driver")
-        os.system('cmd /c "curl https://chromedriver.storage.googleapis.com/101.0.4951.41/chromedriver_win32.zip > '
-                  '{}"'.format(str(os.path.join(folder_driver_path, "chrome_driver.zip"))))
-        with zipfile.ZipFile(os.path.join(folder_driver_path, "chrome_driver.zip"), 'r') as zip_ref:
-            zip_ref.extractall(folder_driver_path)
-        os.remove(os.path.join(folder_driver_path, "chrome_driver.zip"))
+            if chrome_type == "chromedriver_win32" :
+                os.system('cmd /c "curl https://chromedriver.storage.googleapis.com/101.0.4951.41/{}.zip > '
+                  '{}"'.format(chrome_type, (os.path.join(folder_driver_path, "chrome_driver.zip"))))
+                with zipfile.ZipFile(os.path.join(folder_driver_path, "chrome_driver.zip"), 'r') as zip_ref:
+                    zip_ref.extractall(folder_driver_path)
+                os.remove(os.path.join(folder_driver_path, "chrome_driver.zip"))
+
+            else:
+                os.system("wget -N http://chromedriver.storage.googleapis.com/101.0.4951.41/chromedriver_linux64.zip")
+                os.system("unzip chromedriver_linux64.zip")
+                os.system("chmod +x chromedriver")
+                os.system("mv -f chromedriver /usr/local/share/chromedriver")
+                os.system("ln -s /usr/local/share/chromedriver /usr/local/bin/chromedriver")
+                os.system("ln -s /usr/local/share/chromedriver /usr/bin/chromedriver")
+
     else:
         print("Chrome driver exists ! ")
 
@@ -42,11 +67,18 @@ def download_chrome_driver():
 def initialize_driver():
     cwd = os.getcwd()
     driver_path = os.path.join(cwd, "driver", "chromedriver.exe")
-    options = Options()
-    options.headless = False
-    options.add_argument('lang=en')
-    options.add_argument("--window-size=1920,1200")
-    driver = webdriver.Chrome(driver_path, options=options)
+    print(driver_path)
+
+    options = webdriver.ChromeOptions()
+    options.binary_location = "/usr/bin/google-chrome"  # chrome binary location specified here
+    options.add_argument("--start-maximized")  # open Browser in maximized mode
+    options.add_argument('--headless')
+    options.add_argument("--no-sandbox")  # bypass OS security model
+    options.add_argument("--disable-dev-shm-usage")  # overcome limited resource problems
+    options.add_argument('--ignore-ssl-errors=yes')
+    options.add_argument('--ignore-certificate-errors')
+    driver = webdriver.Chrome(r'/usr/local/share/chromedriver', options=options)
+
     return driver
 
 
@@ -70,7 +102,6 @@ def close_error_popup(driver):
 
     except Exception as ex:
         print("error at close_error_popup method : {}".format(ex))
-
 
 
 def scroll_down_first(driver):
